Split every routers and remove event handlers
