+++ fastapi_users/router/__init__.py
@@ -1,7 +1,6 @@
-from fastapi_users.router.common import (  # noqa: F401
-    ErrorCode,
-    Event,
-    EventHandlersRouter,
-)
+from fastapi_users.router.auth import get_auth_router  # noqa: F401
+from fastapi_users.router.common import ErrorCode  # noqa: F401
 from fastapi_users.router.oauth import get_oauth_router  # noqa: F401
+from fastapi_users.router.register import get_register_router  # noqa: F401
+from fastapi_users.router.reset import get_reset_password_router  # noqa: F401
 from fastapi_users.router.users import get_user_router  # noqa: F401
